+++ backend/hatchling/builders/hooks/plugin/interface.py
@@ -30,10 +30,11 @@ class BuildHookInterface(object):  # no cov
     PLUGIN_NAME = ''
     """The name used for selection."""
 
-    def __init__(self, root, config, build_config, directory, target_name, app=None):
+    def __init__(self, root, config, build_config, metadata, directory, target_name, app=None):
         self.__root = root
         self.__config = config
         self.__build_config = build_config
+        self.__metadata = metadata
         self.__directory = directory
         self.__target_name = target_name
         self.__app = app
@@ -77,6 +78,11 @@ class BuildHookInterface(object):  # no cov
             ```
         """
         return self.__config
+
+    @property
+    def metadata(self):
+        # Undocumented for now
+        return self.__metadata
 
     @property
     def build_config(self):